Remove capistrano-rvm dependency
The Infra team found a few years ago that this dependency was no longer needed, and it tended to cause problems (none of which I remember at the moment). For now, the way that capistrano-rvm registers its `rvm:hook` task pre-empts the SSH controlmaster stuff in `dlss-capistrano` 5.x and winds up prompting developers for MFA multiple times before hitting an SSH authentication error. Removing the dependency allows the controlmaster setup to occur and restores the ability to deploy.