Suggest using search
Implement "smart suggest" using the Solr select endpoint rather than
suggest in order to get the document IDs for title results.

When the user types in a string a search for the string plus * of
dct_title_ti and dct_spatial_tmi is performed. The results are stored in
an instance of the Suggestions class which has methods to return
locations, datasets and maps. Each of the suggestions is an instance of
Suggestion which can highlight the query.