Refactor sidebar component and combine secondary portions
This refactors the sidebar component to make it easier to determine
what is being rendered in each of the two sections and to render
it without accidentally duplicating the divider elements.

It also ensures that items that should be grouped together in the
second section are displayed in a single box, as noted in #1437.

Fixes #1437