Defer non-critical CSS without changing cascade order
With 9254fbf, we deferred non-critical CSS by loading it at the
end of the page. However, this had an unintended side-effect:
making the viewer CSS load last in the cascade.

This meant that fixes like https://github.com/geoblacklight/geoblacklight/commit/12bd71fbd36d054353e5047328455966a049b85a
had no effect, because they were loaded as part of geoblacklight's
CSS and then overridden by the viewer's own CSS.

This change switches to the strategy outlined in
https://web.dev/articles/defer-non-critical-css,
which defers CSS loading but reverts to the original DOM order
for the styles, preserving the ability to affect Leaflet's
styles in our own without resorting to !important.

It also removes one instance where we were using !important
to get around this problem.